requirements.txt
affine==2.4.0
albucore==0.0.12
albumentations==1.4.11
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
basemap @ file:///C:/Users/boaky/OneDrive%20-%20Ashesi%20University/schulen/4%20sem%201/Capstone%20Project/App/external_module/basemap-1.4.1-cp39-cp39-win_amd64.whl#sha256=e31eff36677b3970a40591c468eae95c6d67d50df89930b4a7848f723dc60018
basemap-data==1.3.2
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
bqplot==0.12.43
branca==0.7.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
colour==0.1.5
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
earthengine-api==0.1.407
eerepr==0.0.4
efficientnet-pytorch==0.7.1
eval_type_backport==0.2.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
fiona==1.9.6
folium==0.16.0
fonttools==4.53.0
fqdn==1.5.1
fsspec==2024.6.1
future==1.0.0
GDAL @ file:///C:/Users/boaky/OneDrive%20-%20Ashesi%20University/schulen/4%20sem%201/Capstone%20Project/App/external_module/GDAL-3.8.2-cp39-cp39-win_amd64.whl#sha256=376e5d5a5546b3a2d8bbca74d2b050688316375e5a13ea70b0ef61589b632612
geemap @ git+https://github.com/gee-community/geemap@9667bfa8bcc2e4520b33d0a58c249027e97e12df
geocoder==1.38.1
geopandas==0.14.4
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.5
idna==3.7
imageio==2.34.2
importlib_metadata==7.1.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipykernel==6.29.5
ipyleaflet==0.18.2
ipython==8.18.1
ipytree==0.2.2
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
mpmath==1.3.0
munch==4.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
opencv-python==4.10.0.84
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
pretrainedmodels==0.7.4
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
psutil==6.0.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
pyperclip==1.8.2
pyproj==3.6.1
pyshp==2.3.1
PySocks==1.7.1
python-box==7.2.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
rasterio==1.3.10
ratelim==0.1.6
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
scooby==0.10.0
segmentation-models-pytorch==0.3.3
selenium==4.22.0
Send2Trash==1.8.3
shapely==2.0.4
six==1.16.0
sklearn==0.0.post1
smmap==5.0.1
sniffio==1.3.1
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
streamlit==1.35.0
streamlit_folium==0.20.0
sympy==1.13.0
tbb==2021.13.0
tenacity==8.3.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.7.2
timm==0.9.2
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
traittypes==0.2.1
trio==0.26.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.0
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
watchdog==4.0.1
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.11
wsproto==1.2.0
xyzservices==2024.4.0
zipp==3.19.2
